Repair the path to jni.h for libtorch windows build (#93057)
Fixes #86536

It seems like the file is not found when the environment is populate, so the BUILD_JNI flag is false.

To mark it as true, I had to add a `/pytorch/` when adding paths in `POSSIBLE_JAVA_HOMES`. This way, it seems like the file is found and the flag it's true.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/93057
Approved by: https://github.com/malfet, https://github.com/Blackhex