Clarify error message of torch.nn.functional.embedding_bag (#99471)
Fixes #99221  , clarifying the error message to highlight the index from inputs which is responsible for the out-of-bounds error,  while maintaining the reference to the relevant index of offsets as a secondary consideration.

Also takes care of some spelling/grammatical issues with another message (primarily "yout" changed to "your").

Would be happy to iterate on this if there are any issues.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/99471
Approved by: https://github.com/albanD